11-04-2022 - refactor code

+++ src/main/java/com/etiya/rentACar/business/responses/orderedAdditionalServiceResponses/ListOrderedAdditionalServiceDto.java
@@ -11,5 +11,5 @@ public class ListOrderedAdditionalServiceDto {
     private int id;
     private int rentalId;
     private String  additionalServiceName;
-    private int additionalServiceDailyPrice;
+    private double additionalServiceDailyPrice;
 }
